+++ float/src/fbig_cached_ops.rs
@@ -650,7 +650,6 @@ macro_rules! forward_to_fbig {
 impl<R: Round, const B: Word> CachedFBig<R, B> {
     forward_to_fbig!(sqrt);
     forward_to_fbig!(inv);
-    forward_to_fbig!(powi(exp: dashu_int::IBig));
     forward_to_fbig!(sqr);
     forward_to_fbig!(cubic);
 }
@@ -658,6 +657,7 @@ impl<R: Round, const B: Word> CachedFBig<R, B> {
 // Transcendentals that route through the Ziv-backed (or Ziv-dependent) Context methods require
 // `R: ErrorBounds` for their correctness guarantee.
 impl<R: ErrorBounds, const B: Word> CachedFBig<R, B> {
+    forward_to_fbig!(powi(exp: dashu_int::IBig));
     forward_to_context!(ln);
     forward_to_context!(ln_1p);
     forward_to_context!(exp);